ceph.spec.in: ceph-common needs python-argparse unconditionally
The ceph-common subpackage includes the "ceph" command-line interface,
which is implemented as a Python script that does "import argparse".
On newer distros that use Python 2.7, this is no problem because argparse is
included in core. However, on older distros like CentOS 6.6, which uses
Python 2.6, you get an "ImportError: No module named argparse" when you
try to run ceph with just the ceph-common package installed.

http://tracker.ceph.com/issues/12034 Fixes: #12034

Signed-off-by: Nathan Cutler <ncutler@suse.com>